moves pages folder out of components
