+++ src/App.js
@@ -3,10 +3,10 @@ import './App.css';
 import Container from "react-bootstrap/Container";
 import Header from "./components/Header";
 import Footer from "./components/Footer";
-import AboutMe from "./components/pages/AboutMe"
-import Portfolio from "./components/pages/Portfolio"
-import Contact from "./components/pages/Contact"
-import CV from "./components/pages/CV"
+import AboutMe from "./pages/AboutMe"
+import Portfolio from "./pages/Portfolio"
+import Contact from "./pages/Contact"
+import CV from "./pages/CV"
 
 function App() {
 	const [currentPage, setCurrentPage] = useState('AboutMe');
